Ultra mega super last changes............

+++ app/src/main/java/com/bluemeth/simbank/src/domain/AcceptContactRequestUseCase.kt
@@ -11,12 +11,13 @@ class AcceptContactRequestUseCase @Inject constructor(
     private val notificationRepository: NotificationRepository
 ) {
 
-    suspend operator fun invoke(currentUser: User, newContact: User, notification: Notification): Boolean {
+    suspend operator fun invoke(currentUser: User, newContact: User, notification: Notification?): Boolean {
         val contactsUpdated = userRepository.updateUserContacts(currentUser.email, currentUser.contacts)
 
         return if (contactsUpdated) {
             userRepository.updateUserContacts(newContact.email, newContact.contacts)
-            notificationRepository.deleteNotification(notification.id)
+            if(notification != null) notificationRepository.deleteNotification(notification.id)
+            true
         } else {
             false
         }
